Open student search endpoint
