+++ src/main/java/teammates/common/util/Const.java
@@ -763,7 +763,7 @@ public final class Const {
         public static final String SESSION_STATS = "/session/stats";
         public static final String SESSION_SUBMITTED_GIVER_SET = "/session/submitted/giverset";
         public static final String SESSIONS = "/sessions";
-        public static final String STUDENT_ACCOUNTS_SEARCH = "/admin/search/studentaccounts"
+        public static final String STUDENTS_SEARCH = "/admin/search/students";
         public static final String BIN_SESSION = "/bin/session";
         public static final String QUESTIONS = "/questions";
         public static final String QUESTION = "/question";
